refactor(move isInnerLink() to separate file):
