+++ tsconfig.json
@@ -60,8 +60,10 @@
     "src/index.ts",
     "src/install/fetch.ts",
     "src/install/fetchResolution.ts",
+    "src/install/getIsInstallable.ts",
     "src/install/installMultiple.ts",
     "src/install/postInstall.ts",
+    "src/isInnerLink.ts",
     "src/link/index.ts",
     "src/link/linkBins.ts",
     "src/link/linkPeers.ts",
